+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.1.0
 
 Functional Changes:
+- Fix Combat Roll not cancelling attack properly
 - Improve idle animation handling with server side logic
 - Add support `player.entity_interaction_range` attribute, to serve as attack range
 - Add support for weapon attributes resolved from item data component. For example: